1、重写分类前端（treeable-lay插件） 2、添加分类搜索功能

+++ src/main/java/com/example/demo/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.example.demo.controller;
 
+import cn.hutool.core.convert.Convert;
 import com.example.demo.base.Result;
 import com.example.demo.component.RedisService;
 import com.example.demo.jwt.JwtUserDetails;
@@ -8,12 +9,12 @@ import com.example.demo.service.UserService;
 import com.example.demo.dto.UserInfoDto;
 import com.example.demo.dto.RegisterDto;
 import com.example.demo.utils.Constants;
-import com.example.demo.utils.ConvertUtils;
 import com.example.demo.vo.UserInfoVo;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -38,38 +39,50 @@ public class UserController {
     @ApiOperation("注册")
     @PostMapping(value = "/register")
     @ResponseBody
-    public Result register(@Valid RegisterDto registerDto) {
-        User user = userService.register(registerDto);
-        return Result.success(ConvertUtils.convert(user, UserInfoVo.class));
+    public Result<String> register(@Valid RegisterDto registerDto) {
+        int count = userService.register(registerDto);
+        if (count == 1) {
+            return Result.success();
+        } else {
+            return Result.failure();
+        }
     }
 
     @ApiOperation("获取信息")
     @GetMapping("/user")
     @ResponseBody
-    public Result getUserInfo(@ApiIgnore Authentication authentication) {
+    public Result<UserInfoVo> getUserInfo(@ApiIgnore Authentication authentication) {
         JwtUserDetails userDetails = (JwtUserDetails) authentication.getPrincipal();
-        return Result.success(ConvertUtils.convert(userDetails.getUser(), UserInfoVo.class));
+        return Result.success(Convert.convert(UserInfoVo.class, userDetails.getUser()));
     }
 
     @ApiOperation("修改信息")
     @PostMapping("/user")
     @ResponseBody
-    public Result updateUserInfo(@ApiIgnore Principal principal, @Valid UserInfoDto userInfoDto) {
+    public Result<String> updateUserInfo(@ApiIgnore Principal principal, @Valid UserInfoDto userInfoDto) {
         String username = principal.getName(); // SecurityContextHolder上下文
-        userService.updateUserInfoByUsername(username, userInfoDto);
-        redisService.delete(Constants.USER_KEY + username); // 刷新缓存
-        return Result.success();
+        int count = userService.updateUserInfoByUsername(username, userInfoDto);
+        if (count != 0) {
+            redisService.delete(Constants.USER_KEY + username); // 刷新缓存
+            return Result.success();
+        } else {
+            return Result.failure();
+        }
     }
 
     @ApiOperation("修改密码")
     @PostMapping("/user/password")
     @ResponseBody
-    public Result updatePassword(@ApiIgnore Principal principal,
-                                 @RequestParam("password")
-                                 @Size(min = 3, max = 12, message = "密码长度为3-12") String password) {
+    public Result<String> updatePassword(@ApiIgnore Principal principal,
+                                         @RequestParam("password")
+                                         @Size(min = 3, max = 12, message = "密码长度为3-12") String password) {
         String username = principal.getName(); // SecurityContextHolder上下文
-        userService.updatePasswordByUsername(username, password);
-        redisService.delete(Constants.USER_KEY + username); // 刷新缓存
-        return Result.success();
+        int count = userService.updatePasswordByUsername(username, password);
+        if (count != 0) {
+            redisService.delete(Constants.USER_KEY + username); // 刷新缓存
+            return Result.success();
+        } else {
+            return Result.failure();
+        }
     }
 }
